updated compare and upload to use new compare functions

+++ upload.py
@@ -1,10 +1,19 @@
+from skimage.metrics import structural_similarity as ssim
+import numpy as np
 import cv2
 import boto3
 import datetime
 import os
-import numpy as np
 import types
 import os
+import argparse
+
+# Get command line arguments to force run
+parser = argparse.ArgumentParser(description='Take picture and upload')
+parser.add_argument('-f', action='store_true', help='Force run image and upload')
+args = parser.parse_args()
+force = args.f
+
 from dotenv import load_dotenv
 load_dotenv()
 
@@ -16,49 +25,54 @@ today = datetime.datetime.now()
 
 # TODO - add flag for testing: ignore time constraints
 
+def mse(image_a, image_b):
+    # the 'Mean Squared Error' between the two images is the
+	# sum of the squared difference between the two images;
+	# NOTE: the two images must have the same dimension
+    result = np.sum((image_a.astype('float') - image_b.astype('float')) ** 2)
+    result /= float(image_a.shape[0] * image_b.shape[1])
+
+    return result
+
 
 def same_image(original, new):
-    if original.shape == new.shape:
-        print('The images have same size and channels')
-        difference = cv2.subtract(original, new)
+    mse_result = mse(original, new)
+    ssim_result = ssim(original, new)
 
-        b, g, r = cv2.split(difference)
+    # MSE increases the images are less similar
+    # SSIM has smaller values, which indicate less similarity.
 
-        if cv2.countNonZero(b) == 0 and cv2.countNonZero(g) == 0 and cv2.countNonZero(r) == 0:
-            print('The images are completely equal')
-            return True
-
-    return False
+    return ssim_result >= 0.75 and mse_result <= 1000
 
 
-# if today.weekday() <= 4 and 7 < today.hour < 18:
-cam = cv2.VideoCapture(0)
-success, img = cam.read()
-success = True
-duplicate_image = False
+if (today.weekday() <= 4 and 7 < today.hour < 18) or force:
+    cam = cv2.VideoCapture(0)
+    success, img = cam.read()
+    success = True
+    duplicate_image = False
 
-if success:
-    previous_image = cv2.imread('previous.jpg')
+    if success:
+        previous_image = cv2.imread('previous.jpg')
 
-    tstamp = today.strftime('%Y-%m-%d_%H-%M-%S')
-    filename = 'parking-vision-01-img_' + tstamp + '.jpg'
-    print('Saving Image')
-    cv2.imwrite(filename, img)
+        tstamp = today.strftime('%Y-%m-%d_%H-%M-%S')
+        filename = 'parking-vision-01-img_' + tstamp + '.jpg'
+        print('Saving Image')
+        cv2.imwrite(filename, img)
 
-    if previous_image is not None:
-        print('previous image found, comparing')
-        duplicate_image = same_image(cv2.imread(filename), previous_image)
+        if previous_image is not None:
+            print('previous image found, comparing')
+            is_duplicate = same_image(cv2.imread(filename), previous_image)
 
 
-    if duplicate_image is False:
-        print('Images are different, uploading new image')
-        client = boto3.client(
-            's3',
-            aws_access_key_id=AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
-            region_name='us-east-2'
-        )
-        client.upload_file(filename, 'parking-vision', filename)
+        if is_duplicate is False:
+            print('Images are different, uploading new image')
+            client = boto3.client(
+                's3',
+                aws_access_key_id=AWS_ACCESS_KEY_ID,
+                aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
+                region_name='us-east-2'
+            )
+            client.upload_file(filename, 'parking-vision', filename)
 
-        print('renaming image to previous')
-        os.rename(filename, 'previous.jpg')
+            print('renaming image to previous')
+            os.rename(filename, 'previous.jpg')
